feat(web) split right panel and show changes only when nothing selected
